requirements.txt
# nafo-federated-healthcare/requirements.txt
# Install all at once: pip install -r requirements.txt
#
# Phase 1 — Local Training
torch==2.3.0
torchvision==0.18.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4

# Phase 2 — Federation + Privacy
flwr==1.8.0
opacus==1.4.0

# Phase 3 — 5G Digital Twin
simpy==4.1.1

# Phase 4 — Dashboard + Visualisation
streamlit==1.34.0
plotly==5.22.0
umap-learn==0.5.6

# Hospital B — ECG loader
wfdb==4.1.2

# Hospital D — PPG loader
scipy==1.13.0

# Utilities
matplotlib==3.8.4
tqdm==4.66.4